BUG: testing int -> int32 to silence test warninings. See #19138.

+++ numpy/core/tests/test_numeric.py
@@ -3386,7 +3386,7 @@ class TestCross:
     def test_nested_diff_types(self):        
         #gh-19138
         u=np.array([[195,8,9]],np.uint8)
-        v=np.array([250,166,68],np.int)
+        v=np.array([250,166,68],np.int32)
         z=np.array([[950, 11010, -30370]], dtype=np.int32)
         assert_equal(np.cross(v,u), z)
         assert_equal(np.cross(u,v), -z)
